examples/mcard_storage_service/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.1
python-dotenv>=1.0.1
pydantic>=2.6.3
python-multipart>=0.0.9
Pillow>=11.0.0
aiosqlite>=0.19.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
aiohttp>=3.8.0
typing-extensions>=4.5.0
python-dateutil>=2.8.2
mcard-core @ file:///${PWD}/../../
